Built installation profile before site
This allows us to get our own Drush in early which is needed to
run Drush make when installing the rest of Ding2, Drupal and running
Drush commands.

As we are now building the profile in the workspace before copying it
into the site instead of after, we need to make sure that the workspace
is clean.